+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 ### Added
-* New `maximum_speed` parameter to the driving profiles of the directions API to specify speed limit above a certain threshold set in the config file.
+* New `maximum_speed` parameter to the driving profiles of the directions API, for specifying a speed limit, above a certain threshold set in the config file.
 
 ### Fixed
 ### Changed